+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Bug fixes
 
+* [#7871](https://github.com/rubocop-hq/rubocop/pull/7871): Fix an auto-correction bug in `Lint/BooleanSymbol`. ([@knu][])
 * [#7842](https://github.com/rubocop-hq/rubocop/issues/7842): Fix a false positive for `Lint/RaiseException` when raising Exception with explicit namespace. ([@koic][])
 * [#7834](https://github.com/rubocop-hq/rubocop/issues/7834): Fix `Lint/UriRegexp` to register offense with array arguments. ([@tejasbubane][])
 * [#7841](https://github.com/rubocop-hq/rubocop/issues/7841): Fix an error for `Style/TrailingCommaInBlockArgs` when lambda literal (`->`) has multiple arguments. ([@koic][])
@@ -4447,3 +4448,4 @@
 [@yuritomanek]: https://github.com/yuritomanek
 [@egze]: https://github.com/egze
 [@rafaelfranca]: https://github.com/rafaelfranca
+[@knu]: https://github.com/knu